pyproject.toml
[tool.poetry]
name = "drf-course-paper"
version = "0.1.0"
description = ""
authors = ["DenisKhudyakov <denis.hudyakov27@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.6"
djangorestframework = "^3.15.1"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"
requests = "^2.32.3"
ipython = "^8.25.0"
redis = "^5.0.5"
pillow = "^10.3.0"
djangorestframework-simplejwt = "^5.3.1"
django-cors-headers = "^4.3.1"
celery = "^5.4.0"
django-celery-beat = "^2.6.0"
python-telegram-bot = "^21.3"
drf-yasg = "^1.21.7"
coverage = "^7.5.3"


[tool.poetry.group.lint.dependencies]
black = "^24.4.2"
isort = "^5.13.2"
mypy = "^1.10.0"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

